throw error on duplicate codeGen plugin & reuse findReplacementFunc()
